Teach twittering-mode how to detect the current username better.
Previously if the cursor was anywhere other than a real username,
it would treat the current string of text with a contiguous font
face as the username. Now it errors out if the current point is
not on a username. It also treats a user link as a username (so
you can tab to a @name and then hit "n" to jump to that user's
posts)